+++ grid2op/Action/TopologyChangeAction.py
@@ -8,7 +8,14 @@
 
 from grid2op.Action.PlayableAction import PlayableAction
 
+
 class TopologyChangeAction(PlayableAction):
+    """
+    This type of :class:`PlayableAction` implements the modifications of the grid with "change" topological actions.
+
+    It accepts the key words: "change_line_status" and "change_bus".
+    Nothing else is supported and any attempt to use something else will have not impact.
+    """
     def __init__(self, gridobj):
         super().__init__(gridobj)