+++ CHANGELOG.rst
@@ -25,7 +25,7 @@ Change Log
 --------------------
 - [DEPRECATION] `grid2op.Action.TopoAndRedispAction` is now `grid2op.Action.TopologyAndDispatchAction`.
 - [FIXED] a bug that prevented to perform redispatching action if the sum of the action was neglectible (*eg* 1e-14)
-  instead of a bare 0.
+  instead of an exact `0`.
 - [ADDED] `grid2op.Backend.reset` as a way for backends to implement a faster way to reload the grid. Implemented in
   `grid2op.Backend.PandaPowerBackend`
 - [ADDED] `grid2op.Action.TopologySetAction` A subset of actions to limit the agents scope to 'set' operations only
@@ -37,6 +37,9 @@ Change Log
 - [ADDED] `grid2op.Action.DispatchAction` A subset of actions to limit the agents scope to 'redisp' operations only
 - [ADDED] a new method to plot other values that the default one for plotplotly.
 - [ADDED] a better plotting utilities that is now consistent with `PlotPlotly`, `PlotMatplotlib` and `PlotPyGame`
+- [ADDED] a class to replay a logger using `PlotPyGame` class (`grid2op.Plot.EpisodeReplay`)
+- [ADDED] a method to parse back the observations with lower memory footprint and faster, when the observations
+  are serialized into a numpy array by the runner, and only some attributes are necessary.
 
 
 [0.6.1] - 2020-04-??